+++ package.json
@@ -17,6 +17,7 @@
     "knex-db-manager": "^0.5.0",
     "knex-migrate": "^1.7.2",
     "passport": "^0.4.0",
+    "passport-google-oauth20": "^2.0.0",
     "pg": "^7.11.0",
     "sqlite3": "^4.0.8"
   },
